# Improve performance of line rendering