--- a/src/webgl/p5.RendererGL.Retained.js
+++ b/src/webgl/p5.RendererGL.Retained.js
@@ -103,7 +103,9 @@ p5.RendererGL.prototype.createBuffers = function(gId, model) {
     buffers.vertexCount = model.vertices ? model.vertices.length : 0;
   }
 
-  buffers.lineVertexCount = model.lineVertices ? model.lineVertices.length : 0;
+  buffers.lineVertexCount = model.lineVertices
+    ? model.lineVertices.length / 3
+    : 0;
 
   return buffers;
 };
@@ -118,7 +120,7 @@ p5.RendererGL.prototype.drawBuffers = function(gId) {
   const gl = this.GL;
   const geometry = this.retainedMode.geometry[gId];
 
-  if (this._doFill) {
+  if (this._doFill && this.retainedMode.geometry[gId].vertexCount > 0) {
     this._useVertexColor = (geometry.model.vertexColors.length > 0);
     const fillShader = this._getRetainedFillShader();
     this._setFillUniforms(fillShader);
@@ -135,9 +137,6 @@ p5.RendererGL.prototype.drawBuffers = function(gId) {
   }
 
   if (this._doStroke && geometry.lineVertexCount > 0) {
-    const faceCullingEnabled = gl.isEnabled(gl.CULL_FACE);
-    // Prevent strokes from getting removed by culling
-    gl.disable(gl.CULL_FACE);
     const strokeShader = this._getRetainedStrokeShader();
     this._useLineColor = (geometry.model.vertexStrokeColors.length > 0);
     this._setStrokeUniforms(strokeShader);
@@ -146,9 +145,6 @@ p5.RendererGL.prototype.drawBuffers = function(gId) {
     }
     this._applyColorBlend(this.curStrokeColor);
     this._drawArrays(gl.TRIANGLES, gId);
-    if (faceCullingEnabled) {
-      gl.enable(gl.CULL_FACE);
-    }
     strokeShader.unbindShader();
   }