--- a/src/webgl/p5.RendererGL.js
+++ b/src/webgl/p5.RendererGL.js
@@ -1622,14 +1622,10 @@ p5.RendererGL = class RendererGL extends p5.Renderer {
     if (!target) target = this.GL.ARRAY_BUFFER;
     this.GL.bindBuffer(target, buffer);
     if (values !== undefined) {
-      if (values.data) {
-        // bufferData seems to send the whole Float32Array to the GPU even if
-        // it only uses a small part of it, so if we have a DataVector whose size is
-        // way larger than its data, we need to scale it back down
-        values.rescale();
-      }
-      let data = values.data || values;
-      if (!(data instanceof (type || Float32Array))) {
+      let data = values;
+      if (values instanceof p5.DataVector) {
+        data = values.dataArray();
+      } else if (!(data instanceof (type || Float32Array))) {
         data = new (type || Float32Array)(data);
       }
       this.GL.bufferData(target, data, usage || this.GL.STATIC_DRAW);