Addresses https://github.com/processing/p5.js/issues/5975

While the architectural changes in https://github.com/processing/p5.js/pull/6162 might be a bit more difficult to weigh as they are not always better across the board for all browsers, there are some changes that can be pulled out to improve the performance of the existing line rendering architecture without actually changing the way lines are represented.

### Changes
- No longer checks the state of backface culling; instead, it makes the vertex ordering consistent so that we don't need to disable culling when drawing lines
- Removes the need for flattening line data before rendering; instead, it is put into the arrays in a pre-flattened state
- Stores line buffer data in a Float32Array from the beginning so that it can be reused without constantly reallocating new arrays
- Changes `registerEnabled` to a `Set` (not really a perf update, but makes it easier to loop over enabled registers in the future)

### Performance

Test sketch: https://editor.p5js.org/davepagurek/sketches/7jnWpXBq_

#### Many lines
<table>
<tr><th>Browser</th><th>1.6.0</th><th>This PR</th></tr>
<tr><th>Chrome</th><td>60fps</td><td>60fps</td></tr>
<tr><th>Firefox</th><td>15fps</td><td>60fps</td></tr>
<tr><th>Safari</th><td>5fps</td><td>5fps</td></tr>
<tr><th>Mobile Chrome</th><td>6fps</td><td>8fps</td></tr>
</table>

#### Large curve
<table>
<tr><th>Browser</th><th>1.6.0</th><th>This PR</th></tr>
<tr><th>Chrome</th><td>24fps</td><td>28fps</td></tr>
<tr><th>Firefox</th><td>11fps</td><td>20fps</td></tr>
<tr><th>Safari</th><td>20fps</td><td>27fps</td></tr>
<tr><th>Mobile Chrome</th><td>9fps</td><td>14fps</td></tr>
</table>

#### Retained shapes
<table>
<tr><th>Browser</th><th>1.6.0</th><th>This PR</th></tr>
<tr><th>Chrome</th><td>60fps</td><td>60fps</td></tr>
<tr><th>Firefox</th><td>60fps</td><td>60fps</td></tr>
<tr><th>Safari</th><td>60fps</td><td>60fps</td></tr>
<tr><th>Mobile Chrome</th><td>60fps</td><td>60fps</td></tr>
</table>


#### PR Checklist

- [x] `npm run lint` passes
- [x] [Inline documentation] is included / updated
- [x] [Unit tests] are included / updated