@davepagurek 2023-06-22
Backface culling isn't a performance gain on its own, but since some users turn it on occasionally, we used to have some code to temporarily turn it off and then restore its state when drawing lines. It was that saving/restoring bit that was slow, which is gone now 🙂 The culling toggle in the sketch is to help check that lines don't visually break when culling is turned on.